wip: sifnode whitelist output
