+++ smart-contracts/scripts/fetchTokenDetails.js
@@ -9,33 +9,25 @@ const fs = require("fs");
 const axios = require("axios");
 const { ethers } = require("hardhat");
 
-const addressListFile = process.env.ADDRESS_LIST_SOURCE;
-const destinationFolder = "data";
-const destinationFile = generateDestinationFilename();
+const {
+  print,
+  isValidSymbol,
+  generateTodayFilename,
+  generateV1Denom,
+  SIFNODE_MODEL,
+} = require("./helpers/utils");
 
-/**
- * This is what a token looks like at this point to the sifnode team
- * We'll generate a file for them too
- */
-const SIFNODE_MODEL = {
-  is_whitelisted: true,
-  decimals: "",
-  denom: "",
-  base_denom: "",
-  path: "",
-  ibc_channel_id: "",
-  ibc_counterparty_channel_id: "",
-  display_name: "",
-  display_symbol: "",
-  network: "",
-  address: "",
-  external_symbol: "",
-  transfer_limit: "",
-  permissions: ["CLP"],
-  unit_denom: "",
-  ibc_counterparty_denom: "",
-  ibc_counterparty_chain_id: "",
-};
+const addressListFile = process.env.ADDRESS_LIST_SOURCE;
+const destinationFile = generateTodayFilename({
+  directory: "data",
+  prefix: "whitelist_mainnet_update",
+  extension: "json",
+});
+const sifnodeDestinationFile = generateTodayFilename({
+  directory: "data",
+  prefix: "sifnode_mainnet_update",
+  extension: "json",
+});
 
 async function main() {
   print("yellow", "Starting...", true);
@@ -63,7 +55,6 @@ async function main() {
       const name = await instance.name();
       const decimals = await instance.decimals();
       const symbol = await instance.symbol();
-      const v1Denom = generateDenom(symbol);
 
       if (!isValidSymbol(symbol)) {
         print(
@@ -86,10 +77,12 @@ async function main() {
         imageUrl: iconUrl,
       });
 
+      // Now, the sifnode side:
       const sifnodeObj = { ...SIFNODE_MODEL };
-      sifnodeObj.denom = v1Denom;
-      sifnodeObj.base_denom = v1Denom;
+      const v1Denom = generateV1Denom(symbol);
       sifnodeObj.decimals = decimals;
+      sifnodeObj.base_denom = v1Denom;
+      sifnodeObj.denom = v1Denom;
       sifnodeList.push(sifnodeObj);
 
       print(
@@ -111,61 +104,12 @@ async function main() {
   };
 
   fs.writeFileSync(destinationFile, JSON.stringify(output, null, 2));
+  fs.writeFileSync(sifnodeDestinationFile, JSON.stringify(sifnodeList, null, 2));
 
   print("cyan", `DONE! These results have been written to ${destinationFile}:`);
   print("cyan", JSON.stringify(finalList, null, 2));
-}
-
-const colors = {
-  green: "\x1b[42m\x1b[37m",
-  red: "\x1b[41m\x1b[37m",
-  yellow: "\x1b[33m",
-  cyan: "\x1b[36m",
-  close: "\x1b[0m",
-};
-function print(color, message, breakLine) {
-  const lb = breakLine ? "\n" : "";
-  console.log(`${colors[color]}${message}${colors.close}${lb}`);
-}
-
-/**
- * Will return false for a symbol that has spaces and/or special characters in it
- * @param {string} symbol
- * @returns {bool} does the symbol match the RegExp?
- */
-function isValidSymbol(symbol) {
-  const regexp = new RegExp("^[a-zA-Z0-9]+$");
-  return regexp.test(symbol);
-}
-
-function generateDestinationFilename() {
-  // setup month names
-  const monthNames = [
-    "Jan",
-    "Feb",
-    "Mar",
-    "Apr",
-    "May",
-    "Jun",
-    "Jul",
-    "Aug",
-    "Sep",
-    "Oct",
-    "Nov",
-    "Dec",
-  ];
-
-  // get current date (we do it manually so that it's not dependant on user's locale)
-  const today = new Date();
-  const day = String(today.getDate()).padStart(2, "0");
-  const month = monthNames[today.getMonth()];
-  const year = today.getFullYear();
-
-  // transform it in a string with the following format:
-  // whitelist_mainnet_update_14_sep_2021.json
-  const filename = `${destinationFolder}/whitelist_mainnet_update_${day}_${month}_${year}.json`;
-
-  return filename;
+  print('magenta', '---------');
+  print("magenta", `And Sifnode results have been written to ${sifnodeDestinationFile}.`);
 }
 
 async function getTokenMetadata(address) {
@@ -182,11 +126,6 @@ async function getTokenMetadata(address) {
     });
 
   return response?.data?.result?.logo;
-}
-
-function generateV1Denom(symbol) {
-  const denom = "c" + symbol.toLowerCase();
-  return denom;
 }
 
 main()
